feat(portfolio): consolidate holdings management tests and resolve issue #325
- Merged holdings management tests (addAsset, updateAsset, removeAsset,
  updateAssetPrice, updatePortfolioAllocation, runOptimization) from
  test/portfolio/portfolio.service.spec.ts into the main spec file
- Deleted orphaned test/portfolio/portfolio.service.spec.ts (unreachable by
  Jest rootDir=src)
- All 43 tests passing (18 CRUD + 25 holdings/optimization)

Closes #325
